pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.0",
    "langchain-groq>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
]
